remove 2032 browser.migrate.automigrate.enabled
This is a pref to hide a UI doorhanger which allows an end user to KEEP or UNDO auto-migrated "stuff" from another browser (eg from IE etc). This is part of funnelweb, which is an upcoming makeover of the "onboarding experience" (to retain users in a new FF) - this includes auto-importing their bookmarks etc from another browser in the background and THEN asking if they want to keep them or undo. We cannot block the UI to this for obvious reasons. Funnelweb also includes this like a new Tour UI, activity streams etc

https://bugzilla.mozilla.org/show_bug.cgi?id=1297479
https://bugzilla.mozilla.org/show_bug.cgi?id=1322718#c1

Tour thingy: https://www.ghacks.net/2017/03/31/firefox-photon-new-design-mockups-show-interface-and-more/
Activity Streams: https://www.ghacks.net/2017/04/09/firefox-57-photon-mockups-activity-stream-library-compact-mode-more/